Update for JRuby 9.2+
These are minimal changes to support working on JRuby 9.2 and up,
which are the currently-supported JRuby versions. I was unable to
run the tests due to errors building (something in the toolchain
is using an http URL for Maven that no longer exists).

* CallBlock.newCallClosure with Arity was deprecated some time ago
  and removed in 9.3. The Signature version works on 9.2+.
* RuntimeHelpers is replaced with Helpers.
* createNativeRaiseException is deprecated, so the patch converts
  IOException instances to their equivalent Ruby IOError subtypes.

There may be more work needed to get this fully-functional on
JRuby 9.2+ but this is at least a minimum set of changes.